backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.30.0
attrs==25.4.0
bcrypt==5.0.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.3
databases==0.9.0
ecdsa==0.19.1
fastapi==0.119.1
frozenlist==1.8.0
greenlet==3.2.4
h11==0.16.0
idna==3.11
iniconfig==2.3.0
iso8601==2.1.0
multidict==6.7.0
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
propcache==0.4.1
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
pypika-tortoise==0.6.2
pytest==8.4.2
python-dotenv==1.1.1
python-jose==3.5.0
pytz==2025.2
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
tortoise-orm==0.25.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
yarl==1.22.0
